Starting to create the design of the site.
First added index.html and a css page. The design is mimicking the existing site, but from there I'm going to do custom design to suit the pupose of this site as an internal managment app. I will try to maintain as much consistency as I can with the L&F of the public site.